test(ai): split diplomatic scoring suppression tests for line limit
Refs #2509
